+++ Sampling/random_crop_data.py
@@ -7,27 +7,30 @@ import os
 import numpy as np
 import sys
 import time
+import shutil
 
 # ------ configuration ------
 
 scannet_dir = "/home/dtc/Data/ScanNet"
 
+processing_time_only = True  # whether only care about processing time
+
 # keep ratios
-keep_ratio_arr = range(2, 101, 2)
+keep_ratio_arr = range(2, 101, 10)
 
 # --- end of configuration
 
 data_type = "Random"
 
+# clear tmp
+files = glob.glob("../tmp/*")
+for file in files:
+    os.remove(file)
+
 
 def crop_data(keep_ratio):
 
     start_time = time.time()
-
-    # clear tmp
-    files = glob.glob("../tmp/*")
-    for file in files:
-        os.remove(file)
 
     original_dir = os.path.join(scannet_dir, "Pth/Original")
     dst_dir = os.path.join(scannet_dir, "Pth/{}".format(data_type))
@@ -48,7 +51,7 @@ def crop_data(keep_ratio):
         original_data.ravel().tofile(tmp_file_name)
 
         mycmd = "../Cpp/sample_data/build/sample_data {} {} {} {}"\
-            .format(data_type.lower(), dst_dir, src_filename, keep_ratio)
+            .format(data_type.lower(), tmp_dir, src_filename, keep_ratio)
         os.system(mycmd)
         os.remove(tmp_file_name)
 
@@ -56,40 +59,48 @@ def crop_data(keep_ratio):
         if not os.path.exists(src_trim_file):
             sys.exit("Error, file " + src_trim_file + " does not exist")
 
-        # calculate number of points
-        new_data = np.fromfile(src_trim_file, "<f4")
-        new_data = np.reshape(new_data, (-1, 7))
-        new_coords = new_data[:, :3]
-        new_colors = new_data[:, 3:6]
-        new_labels = new_data[:, 6]
+        # # calculate number of points
+        # new_data = np.fromfile(src_trim_file, "<f4")
+        # new_data = np.reshape(new_data, (-1, 7))
+        # new_coords = new_data[:, :3]
+        # new_colors = new_data[:, 3:6]
+        # new_labels = new_data[:, 6]
 
-    if not os.path.exists(os.path.join(dst_dir, "{}".format(keep_ratio))):
-        os.makedirs(os.path.join(dst_dir, "{}".format(keep_ratio)))
+    if not processing_time_only:
+        if not os.path.exists(os.path.join(dst_dir, "{}".format(keep_ratio))):
+            os.makedirs(os.path.join(dst_dir, "{}".format(keep_ratio)))
 
-    # copy files to dst
-    files = sorted(glob.glob(os.path.join(tmp_dir, '*.trim')))
-    for trim_file in files:
-        src_filename = os.path.basename(trim_file)
-        src_filename = src_filename[:-5]  # remove .trim
-        new_data = np.fromfile(trim_file, "<f4")
-        new_data = np.reshape(new_data, (-1, 7))
-        os.remove(trim_file)
+        # copy files to dst
+        files = sorted(glob.glob(os.path.join(tmp_dir, '*.trim')))
+        for trim_file in files:
+            src_filename = os.path.basename(trim_file)
+            src_filename = src_filename[:-5]  # remove .trim
+            new_data = np.fromfile(trim_file, "<f4")
+            new_data = np.reshape(new_data, (-1, 7))
+            os.remove(trim_file)
 
-        new_coords = new_data[:, :3]
-        new_colors = new_data[:, 3:6]
-        new_labels = new_data[:, 6]
+            new_coords = new_data[:, :3]
+            new_colors = new_data[:, 3:6]
+            new_labels = new_data[:, 6]
 
-        new_coords = np.array(new_coords, "float32")
-        new_colors = np.array(new_colors, "float32")
-        new_labels = np.array(new_labels, "float32")
+            new_coords = np.array(new_coords, "float32")
+            new_colors = np.array(new_colors, "float32")
+            new_labels = np.array(new_labels, "float32")
 
-        dst_file_path = os.path.join(dst_dir, "{}/{}".format(keep_ratio, src_filename))
-        # print(trim_file, " ---> ", dst_file_path)
-        new_coords = np.ascontiguousarray(new_coords)
-        new_colors = np.ascontiguousarray(new_colors)
-        new_labels = np.ascontiguousarray(new_labels)
-        torch.save((new_coords, new_colors, new_labels), dst_file_path)
+            dst_file_path = os.path.join(dst_dir, "{}/{}".format(keep_ratio, src_filename))
+            # print(trim_file, " ---> ", dst_file_path)
+            new_coords = np.ascontiguousarray(new_coords)
+            new_colors = np.ascontiguousarray(new_colors)
+            new_labels = np.ascontiguousarray(new_labels)
+            torch.save((new_coords, new_colors, new_labels), dst_file_path)
 
+    shutil.move(os.path.join(tmp_dir, "time.txt"), os.path.join(tmp_dir, "time.txt.{}".format(keep_ratio)))
+    # clear tmp
+    files = glob.glob("../tmp/*")
+    for file in files:
+        if os.path.basename(file).startswith("time.txt."):
+            continue
+        os.remove(file)
     print("------ ratio {}%, {:.2f}s".format(keep_ratio, time.time() - start_time))
